Move schema to synchronizer folder
