+++ framework/src/application/node/node.js
@@ -169,6 +169,10 @@ module.exports = class Node {
 
 			this._subscribeToEvents();
 
+			this.channel.subscribe('app:networkReady', async () => {
+				await this._startLoader();
+			});
+
 			this.channel.subscribe('app:ready', async () => {
 				await this._startForging();
 			});
@@ -447,6 +451,10 @@ module.exports = class Node {
 		this.modules.transport = this.transport;
 		this.modules.bft = this.bft;
 		this.modules.synchronizer = this.synchronizer;
+	}
+
+	async _startLoader() {
+		return this.synchronizer.loadUnconfirmedTransactions();
 	}
 
 	async _forgingTask() {
